[None][test] Rename attention backend test modules
